requirements.txt
pandas
requests
python-bitcoinrpc